phase3: fix five review issues — dead TLS, int64 overflow, validation sampling, AVX2 hsum, thread kernel propagation
Issue #12: remove g_deterministic_seed_tls (declared, never read in decode path).
Issue #9: max_dense_gradient_elements int→int64_t in DBSOptionsC (dbs.h, .cpp, ctypes);
  adds _pad0 field in Python struct to match C alignment; fixes silent 2 GB overflow for
  large T×K×V gradient tensors.
Issue #4: validate_logprob_tensor now samples ≤1000 evenly-spaced elements instead of
  scanning the full O(T×K×V) tensor, keeping validation O(1) in input size.
Issue #10: AVX2 hsum256 replaced with _mm256_hadd_ps two-stage horizontal reduction,
  eliminating the 32-byte aligned store + scalar loop.
Issue #7: dbs_decode_batch and dbs_decode_batch_variable capture g_kernel_override_enabled
  / g_kernel_override from the calling thread and apply them inside each worker thread,
  so kernel-override tests propagate correctly to the thread pool.

All 43 tests pass (1 skipped, unchanged).

+++ include/dbs.h
@@ -38,7 +38,7 @@ typedef struct DBSOptionsC {
     int soft_topk_max_iters;
     int min_length;
     int validate_inputs;
-    int max_dense_gradient_elements;
+    int64_t max_dense_gradient_elements;
     int reserved0;
     int reserved1;
 } DBSOptionsC;
